Code restructuring - Create lib/prefutils.py
Move main part of code outside of default.py
According to code quality/sizing rule.
